Eduardo

+++ Models/Format.cs
@@ -1,6 +1,6 @@
 namespace AicaDocsApi.Models;
 
-public enum Format : byte
+public enum Format : short
 {
     Pdf,
     Word
